+++ include/wex/plot/plscatterplot.h
@@ -12,7 +12,7 @@ public:
 	wxPLScatterPlot( const std::vector<wxRealPoint> &data,
 		const wxString &label = wxEmptyString,
 		const wxColour &col = *wxBLUE,
-		int size = 1,
+		double size = 1,
 		bool scale = false );
 
 	virtual ~wxPLScatterPlot();
@@ -29,12 +29,12 @@ public:
 	virtual void DrawInLegend( wxPLOutputDevice &dc, const wxPLRealRect &rct);
 
 	void SetColour( const wxColour &col ) { m_colour = col; }
-	void SetSize( int radius ) { m_radius = radius; }
+	void SetSize( double radius ) { m_radius = radius; }
 	void SetLineOfPerfectAgreementFlag(bool flagValue);
 
 protected:
 	wxColour m_colour;
-	int m_radius;
+	double m_radius;
 	bool m_scale;
 	bool m_drawLineOfPerfectAgreement;
 	bool m_isLineOfPerfectAgreementDrawn;
